registration form, update post's tags

+++ myblog/views.py
@@ -8,12 +8,13 @@ from django.http import Http404
 from django.views import generic
 from django.contrib.auth import authenticate, login, logout
 from django.core.urlresolvers import reverse
-from .forms import PostForm, CommentForm
+from .forms import PostForm, CommentForm, UserForm
 from django.contrib.auth.models import User
 from django.core import serializers
-# from django.contrib.auth.decorators import login_required
+from django.contrib.auth.decorators import login_required
 import json
 from collections import Counter
+from django.contrib.auth.hashers import make_password
 
 def index(request):
     posts=Post.objects.filter(is_published=True).order_by('-pub_date')[:6]
@@ -30,7 +31,7 @@ def signin(request):
             return HttpResponseRedirect(reverse('index'))
             # return HttpResponse('wellcome')
         else:
-            return HttpResponse('not active :(')
+            return render(request,'myblog/inactive.html')
     else:
         return HttpResponse('incorrect!!')
 
@@ -38,37 +39,45 @@ def log_out(request):
     logout(request)
     return HttpResponseRedirect(reverse('index'))
 
-# @login_required
+# @login_required(login_url='/blog/login/')
 def create_post(request):
-    if request.user.is_authenticated(): # uathentication
-        if request.method == 'POST':
-            # return HttpResponse(request.POST['is_published'])
-            form=PostForm(request.POST,request.FILES)
-            if form.is_valid():
-                new_post=form.save(commit=False)
-                new_post.user_id=request.user.id
-                if request.POST.has_key('is_published'):
-                    new_post.pub_date=datetime.datetime.now()
-                tags=[ tag.lower() for tag in request.POST['tags'].split(',')] #for case insensitive comparison
-                tags=list(set(tags)) #unique list
-                new_tags=[Tag(name=tag) for tag in tags if not Tag.objects.filter(name__iexact=tag).exists() ]
-                Tag.objects.bulk_create(new_tags)
-                new_post.save()
-                for tag in tags:
-                    new_post.tag_set.add(Tag.objects.get(name=tag))
-                return HttpResponseRedirect(reverse('create_post'))
+    # if request.user.is_authenticated(): # uathentication
+        if request.user.is_active:
+            if request.method == 'POST':
+                # return HttpResponse(request.POST['is_published'])
+                form=PostForm(request.POST,request.FILES)
+                if form.is_valid():
+                    new_post=form.save(commit=False)
+                    new_post.user_id=request.user.id
+                    if request.POST.has_key('is_published'):
+                        new_post.pub_date=datetime.datetime.now()
+                    tags=[ tag.lower() for tag in request.POST['tags'].split(',')] #for case insensitive comparison
+                    tags=list(set(tags)) #unique list
+                    new_tags=[Tag(name=tag) for tag in tags if not Tag.objects.filter(name__iexact=tag).exists() ]
+                    Tag.objects.bulk_create(new_tags)
+                    new_post.save()
+                    for tag in tags:
+                        new_post.tag_set.add(Tag.objects.get(name=tag))
+                    return HttpResponseRedirect(reverse('create_post'))
 
-        form = PostForm(request.POST or None)
-        all_tags=json.dumps([unicode(i) for i in Tag.objects.values_list('name',flat=True)])
+            form = PostForm(request.POST or None)
 
-        return render(request,'myblog/create-post.html',{'form':form,'tags':all_tags})
-    else:
-        return render(request,'myblog/login.html')
+            all_tags=json.dumps([unicode(i) for i in Tag.objects.values_list('name',flat=True)])
+
+            return render(request,'myblog/create-post.html',{'form':form,'tags':all_tags})
+        else:
+            return render(request,'myblog/inactive.html')
+    # else:
+    #     return render(request,'myblog/login.html')
     # return render(request,'myblog/create-post.html')
 
+@login_required(login_url='/blog/login/')
 def show_posts(request):
-    posts=Post.objects.filter(user_id=request.user.id).order_by('-pub_date')
-    return render(request,'myblog/show-posts.html',{'posts':posts})
+    if request.user.is_active:
+        posts=Post.objects.filter(user_id=request.user.id).order_by('-pub_date')
+        return render(request,'myblog/show-posts.html',{'posts':posts})
+    else:
+        return render(request,'myblog/inactive.html')
 
 # def single_post(request,pk):
 #     post=get_object_or_404(Post,pk=pk)
@@ -94,23 +103,46 @@ class PostView(generic.DetailView): #single post page
         context['similars']=[tuple[0] for tuple in Counter(Post.objects.filter(is_published=True,tag__name__in=['php','python'])).most_common(4)]
         return context
 
+@login_required(login_url='/blog/login/')
 def update_post(request,pk):
-    if request.method == 'POST':
-        post=get_object_or_404(Post,id=pk)
-        form = PostForm(request.POST,request.FILES,instance=post)
-        if form.is_valid():
-            new_post=form.save(commit=False)
-            if request.POST.has_key('is_published'):
-                new_post.pub_date=datetime.datetime.now()
-            new_post.save()
-        return HttpResponseRedirect(reverse('update_post',args=[pk]))
+    if request.user.is_active:
+        if request.method == 'POST':
+            post=get_object_or_404(Post,id=pk)
+            form = PostForm(request.POST,request.FILES,instance=post)
+            if form.is_valid():
+                new_post=form.save(commit=False)
+                if request.POST.has_key('is_published'):
+                    new_post.pub_date=datetime.datetime.now()
 
-    form=PostForm(instance=get_object_or_404(Post,pk=pk))
-    return render(request,'myblog/update-post.html',{'form':form,'id':pk})
+                new_post.tag_set.clear()
+                new_tags_list=list(set([ tag.lower() for tag in request.POST['tags'].split(',')])) #for case insensitive comparison
+                new_tags=[Tag(name=tag) for tag in new_tags_list ]
+                new_new_tags=[tag for tag in new_tags if not Tag.objects.filter(name__iexact=tag.name).exists() ]
+                Tag.objects.bulk_create(new_new_tags)
+                for tag in new_tags:
+                    new_post.tag_set.add(Tag.objects.get(name=tag.name))
+                new_post.save()
 
+                return HttpResponseRedirect(reverse('show_posts'))
+            else:
+                all_tags=json.dumps([unicode(i) for i in Tag.objects.values_list('name',flat=True)])
+                old_tags= str(','.join(Tag.objects.filter(post=get_object_or_404(Post,id=pk)).values_list('name',flat=True)))
+                return render(request,'myblog/update-post.html',{'form':form,'id':pk,'tags':all_tags,'old_tags':old_tags})
+
+        form=PostForm(instance=get_object_or_404(Post,pk=pk))
+        all_tags=json.dumps([unicode(i) for i in Tag.objects.values_list('name',flat=True)])
+        old_tags= str(','.join(Tag.objects.filter(post=get_object_or_404(Post,id=pk)).values_list('name',flat=True)))
+        return render(request,'myblog/update-post.html',{'form':form,'id':pk,'tags':all_tags,'old_tags':old_tags})
+    else:
+        return render(request,'myblog/inactive.html')
+
+@login_required(login_url='/blog/login/')
 def delete_post(request,pk):
-    post=Post.objects.get(pk=pk).delete()
-    return HttpResponseRedirect(reverse('show_posts'))
+    if request.user.is_active:
+        Post.objects.get(pk=pk).delete()
+        return HttpResponseRedirect(reverse('show_posts'))
+    else:
+        return render(request,'myblog/inactive.html')
 
 # def create_comment(request):
 #     form = CommentForm(request.POST)
@@ -124,35 +156,53 @@ def delete_post(request,pk):
 #         new_comment.save()
 #         return  HttpResponse('done')
 
+@login_required(login_url='/blog/login/')
 def create_comment(request):
-    new_comment=Comment(content=request.POST['content'],)
-    if request.POST['parent'] != 'c':
-         new_comment.parent=get_object_or_404(Comment,id=request.POST['parent'])
+    if request.user.is_active:
+        new_comment=Comment(content=request.POST['content'],)
+        if request.POST['parent'] != 'c':
+             new_comment.parent=get_object_or_404(Comment,id=request.POST['parent'])
+        else:
+            new_comment.parent=None
+        new_comment.user=get_object_or_404(User,id=request.POST['user'])
+        new_comment.post=get_object_or_404(Post,id=request.POST['post'])
+        new_comment.save()
+        return HttpResponse(serializers.serialize('json',Comment.objects.filter(id=new_comment.id)))
     else:
-        new_comment.parent=None
-    new_comment.user=get_object_or_404(User,id=request.POST['user'])
-    new_comment.post=get_object_or_404(Post,id=request.POST['post'])
-    new_comment.save()
-    return HttpResponse(serializers.serialize('json',Comment.objects.filter(id=new_comment.id)))
+        return render(request,'myblog/inactive.html')
 
+@login_required(login_url='/blog/login/')
 def delete_comment(request):
-    comment=get_object_or_404(Comment,id=request.POST['id'])
-    comment.delete()
-    return HttpResponse('done')
+    # if request.user.is_active:
+        comment=get_object_or_404(Comment,id=request.POST['id'])
+        comment.delete()
+        return HttpResponse('done')
+    # else:
+    #     return render(request,'myblog/inactive.html')
 
+@login_required(login_url='/blog/login/')
 def like_comment(request):
-    comment=get_object_or_404(Comment,id=request.POST['comment_id'])
-    user=request.POST['user_id']
-    comment.likes.add(user)
-    return HttpResponse('done')
+    # if request.user.is_active:
+        comment=get_object_or_404(Comment,id=request.POST['comment_id'])
+        user=request.POST['user_id']
+        comment.likes.add(user)
+        return HttpResponse('done')
+    # else:
+    #     return render(request,'myblog/inactive.html')
 
+@login_required(login_url='/blog/login/')
 def unlike_comment(request):
-    comment=get_object_or_404(Comment,id=request.POST['comment_id'])
-    user=request.POST['user_id']
-    comment.likes.remove(user)
-    return HttpResponse('done')
+    # if request.user.is_active:
+        comment=get_object_or_404(Comment,id=request.POST['comment_id'])
+        user=request.POST['user_id']
+        comment.likes.remove(user)
+        return HttpResponse('done')
+    # else:
+    #     return render(request,'myblog/inactive.html')
+
 def get_tags(request):
     return HttpResponse(serializers.serialize('json',Tag.objects.all().values_list('name',flat=True)))
+
 
 def similar_posts(request,tag_id):
     posts=Post.objects.filter(tag__id=tag_id,is_published=True).order_by('-pub_date')
@@ -170,19 +220,42 @@ def similar_posts(request,tag_id):
 
 
     # return HttpResponseRedirect(reverse('PostView'))
+@login_required(login_url='/blog/login/')
 def mark_post(request,post_id):
-    post=get_object_or_404(Post,id=post_id)
-    post.marked.add(get_object_or_404(User,id=request.POST['user_id']))
-    return HttpResponse('done')
+    # if request.user.is_active:
+        post=get_object_or_404(Post,id=post_id)
+        post.marked.add(get_object_or_404(User,id=request.POST['user_id']))
+        return HttpResponse('done')
+    # else:
+    #     return render(request,'myblog/inactive.html')
 
+@login_required(login_url='/blog/login/')
 def unmark_post(request,post_id):
-    post=get_object_or_404(Post,id=post_id)
-    post.marked.remove(request.POST['user_id'])
-    return HttpResponse('done')
+    # if request.user.is_active:
+        post=get_object_or_404(Post,id=post_id)
+        post.marked.remove(request.POST['user_id'])
+        return HttpResponse('done')
+    # else:
+    #     return render(request,'myblog/inactive.html')
 
+@login_required(login_url='/blog/login/')
 def marked_posts(request):
-    posts=request.user.mark.all()
-    return render(request,'myblog/marked-posts.html',{'posts':posts})
+    if request.user.is_active:
+        posts=request.user.mark.all()
+        return render(request,'myblog/marked-posts.html',{'posts':posts})
+    else:
+        return render(request,'myblog/inactive.html')
+
+def signup(request):
+    if request.method == 'POST':
+        form=UserForm(request.POST)
+        if form.is_valid():
+            user=form.save(commit=False)
+            user.password=make_password(user.password)
+            user.save()
+            return render(request,'myblog/login.html')
+    form=UserForm(request.POST or None)
+    return render(request,'myblog/signup.html',{'form':form})
 
 # def post(request,post_id):
 #
